+++ tests/Support/StubErrorHandler.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Files\Tests\Support;
+
+final class StubErrorHandler
+{
+    public function __invoke(): bool
+    {
+        return false;
+    }
+}