+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Enh #72: In `PathMatcher::only()` and `PathMatcher::except()` methods move a type hints from annotation
   to signature (@xepozz)
 - Enh #87: Use FQN for built-in PHP functions, minor refactoring `FileHelper` and `PathMatcher` (@Tigrov)
+- Bug #91: Restore error handler after handling exception (@vjik)
 
 ## 2.0.0 July 05, 2022